+++ internal/encryption/base.go
@@ -95,10 +95,6 @@ func (s *baseEncryption) encrypt(data []byte, enhance func(basedata) interface{}
 	encryptor := s.encMethods[0]
 
 	if unencrypted.Is(encryptor) {
-		// ensure that the method is defined when Enforced is true
-		if s.enforced {
-			return nil, fmt.Errorf("unable to use unencrypted method for %q when enforced = true", s.name)
-		}
 		return data, nil
 	}
 
@@ -143,9 +139,6 @@ func (s *baseEncryption) decrypt(data []byte, validator func([]byte) error) ([]b
 		// Yep, it's already decrypted
 		for _, method := range s.encMethods {
 			if unencrypted.Is(method) {
-				if s.enforced {
-					return nil, fmt.Errorf("unable to use unencrypted method when enforced = true")
-				}
 				return data, nil
 			}
 		}